Shop menu is finally starting to take shape. I think the ShopKeeper class is done now though. classes/containers inventory.hxx: Added a struct for the Inventory Menu entries since I didn't like the way std::tuple looked/it was overkill. shop_inventory.hxx: Only a single new method atm classes/fighters/player.hxx: for now it's a friend of the ShopKeeper class. Will fix it by adding new getters later.
menus/input.hxx: Made the text-wrap function replace the character we're searching for with a space to make it inline with how I want it to work.
menus/shop.hxx: The menu is starting to take shape.

tests/shop_demo.cpp: The shop menu is starting to take shape finally.
tests/shop_test.cpp: This is the actual shop test file that I'll be using later.

CMakeLists.txt: Added the new files to it's list.