Made it so that the battle menu works and that the player can be rewarded items. Also created the LootTable class.
classes/containers
inventory.hxx: Added a method to get the number of items it currently contains.
loot_table.hxx: A new sub class for mobs with the probabilities of items being dropped.

classes/fighters
actor.hxx: Made the unequip weapon/armor to have it actually remove it from their equipment and put it back in inventory.
mob.hxx: Made it so that the mobs returned a rewards struct with items and the like in it.

classes/items/itembase.hxx: Just added the tier information in there. Will have to add scaling for tier later on.

menus/battle.hxx: Made it so that it uses the new struct I made and rewards the player accordingly.

tests/battle_demo.cpp: Made the mobs have some items to test out the new feature.

CMakeLists.txt: Added the new files to it.

terminal.hxx: Added a new function to move the terminal to a certain row and then clear from there on to the bottom.